backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
cloudinary
